# docs/api/OPENAPI.yaml
openapi: 3.1.0
info:
  title: Automation Opportunity Scanner API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /projects/{projectId}:
    parameters:
      - in: path
        name: projectId
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Read project with its processes
      responses:
        '200': { description: OK }
        '404': { description: Not found in tenant scope }
    patch:
      summary: Rename or archive project (archive requires owner/admin)
      responses:
        '200': { description: Updated }
        '403': { description: Role not permitted }
  /projects/{projectId}/processes:
    post:
      summary: Create process
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Created }
  /processes/{processId}:
    parameters:
      - in: path
        name: processId
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Read process with version history
      responses:
        '200': { description: OK }
        '404': { description: Not found in tenant scope }
  /processes/{processId}/intake:
    post:
      summary: Save workflow intake (new draft version; prior drafts superseded)
      parameters:
        - in: path
          name: processId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Saved }
...
  /processes/{processId}/documents:
    parameters:
      - in: path
        name: processId
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: List sanitized metadata for process documents
      responses:
        '200': { description: Document metadata }
    post:
      summary: Upload UTF-8 text or a readable PDF as untrusted process evidence
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201': { description: Document evidence accepted }
        '413': { description: File too large }
        '415': { description: Unsupported media type }
...